refactor(error): move to `error` folder
