+++ src/callback/type/error-callback.type.ts
@@ -1,6 +1,6 @@
 // External import: Type.
 import { ResultCallback } from '@angular-package/type';
 // Import: Type.
-import { ErrorType } from './error-type.type';
+import { ErrorType } from '../../error/type/error-type.type';
 // Export: Type.
 export type ErrorCallback = (message: string, type?: ErrorType, on?: boolean) => ResultCallback;
