+++ src/error/type/error-type.type.ts
